LayoutTests/platform/chromium/fast/dom/Geolocation/callback-exception-expected.txt

CONSOLE MESSAGE: line 25: Uncaught Error: Exception in success callback
Tests that when an exception is thrown in the success callback, the error callback is not invoked. Note that this test throws an exception which is not caught.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy
PASS successfullyParsed is true

TEST COMPLETE
